- hltypes update for WinRT

+++ src/rendersystems/DirectX/11/DirectX11_VertexShader.h
@@ -43,7 +43,7 @@ namespace april
 	protected:
 		ComPtr<ID3D11VertexShader> dx11Shader;
 		unsigned char* shaderData;
-		long shaderSize;
+		int shaderSize;
 
 	};
